+++ R/get.R
@@ -1,9 +1,11 @@
 #' Obtain the knots from a broken stick model
 #'
 #' @param object An object of class \code{brokenstick}
-#' @param what A character vector of length 1. Valid values are
-#' \code{"all"}, \code{"internal"}, \code{"boundary"}, \code{"dropfirst"}
-#' or \code{"droplast"}. The default is \code{what = "all"}.
+#' @param whatknots A character vector of length 1 specifies the knot set.
+#' Valid values are \code{"all"}, \code{"internal"}, \code{"boundary"},
+#' \code{"dropfirst"} and \code{"droplast"}. The default is
+#' \code{whatknots = "all"}
+#' @param what Deprecated. Use `whatknots` instead.
 #' @return A vector with knot locations, either both, internal only or
 #' boundary only. The result is \code{NULL} if \code{object} does not
 #' have proper class. Returns \code{numeric(0)} if
@@ -12,22 +14,27 @@
 #' get_knots(fit_200, "internal")
 #' @export
 get_knots <- function(object,
-                      what = c("all", "internal", "boundary", "dropfirst", "droplast")) {
+                      whatknots = c("all", "internal", "boundary", "dropfirst", "droplast"),
+                      what = "all") {
   stopifnot(inherits(object, c("brokenstick")))
-
-  what <- match.arg(what)
+  if (!missing(what)) {
+    warning("argument what is deprecated; please use whatknots instead.",
+            call. = FALSE)
+    whatknots <- what
+  }
+  whatknots <- match.arg(whatknots)
   internal <- object$internal
   # legacy for objects created before v2.0
   if (is.null(internal)) internal <- object$knots
   boundary <- object$boundary
   internal <- internal[internal > boundary[1L] & internal < boundary[2L]]
 
-  result <- switch(what,
-    all = c(boundary[1L], internal, boundary[2L]),
-    internal = internal,
-    boundary = boundary,
-    dropfirst = c(internal, boundary[2L]),
-    droplast = c(boundary[1L], internal)
+  result <- switch(whatknots,
+                   all = c(boundary[1L], internal, boundary[2L]),
+                   internal = internal,
+                   boundary = boundary,
+                   dropfirst = c(internal, boundary[2L]),
+                   droplast = c(boundary[1L], internal)
   )
   return(result)
 }
@@ -62,24 +69,36 @@ get_r2 <- function(object, newdata = NULL) {
 #' @param names  A vector of column names of. If not specified, the function
 #'  automatically drops the entries corresponding to the right boundary. Specify
 #'  `names = "all"` to prevent dropping.
+#' @inheritParams get_knots
 #' @return A numeric matrix, possibly with zero rows and columns if no names match
 #' @examples
 #' f1 <- brokenstick(hgt_z ~ age | id, smocc_200[1:1000, ], knots = 0:2, seed = 1)
-#' get_omega(f1, "cor", c("age_1", "age_2"))
+#' get_omega(f1, what = "cor", names = c("age_1", "age_2"))
 #' @export
-get_omega <- function(x, what = c("cov", "cor"), names = NULL) {
+get_omega <- function(x,
+                      what = c("cov", "cor"),
+                      whatknots = c("all", "internal", "boundary", "dropfirst", "droplast"),
+                      names = NULL) {
   stopifnot(inherits(x, "brokenstick"))
   what <- match.arg(what)
+  whatknots <- match.arg(whatknots)
+
   omega <- x$omega
+  v <- colnames(omega)
+  nameset <- switch(whatknots,
+                    all = v,
+                    internal = v[c(-1L, -length(v))],
+                    boundary = v[c(1L, length(v))],
+                    dropfirst = v[-1L],
+                    droplast = v[-length(v)])
+  if (!is.null(names)) {
+    nameset <- intersect(names, v)
+  }
   if (length(names) == 1L && names == "all") {
-    names <- colnames(omega)
+    nameset <- v
   }
-  if (is.null(names)) {
-    names <- colnames(omega)[1L:length(colnames(omega)) - 1L]
-  }
-  names <- intersect(names, colnames(omega))
-  if (length(names)) {
-    omega <- omega[names, names, drop = FALSE]
+  if (length(nameset)) {
+    omega <- omega[nameset, nameset, drop = FALSE]
   } else {
     omega <- matrix(NA_real_, 0L, 0L)
   }