+++ NEWS.md
@@ -1,5 +1,7 @@
 # brokenstick 2.2.1
 
+- Replaces `what` argument in `get_knots()` to `whatknots`
+- Adds support for `whatknots` in `get_omega()`
 - Extends capabilities of `plot_trajectory()` with `shape` and `linetype` options
 - Adds an example to `plot.brokenstick()` on how to create a decent black and white figure of trajectories
 - Replaces `knots = 0:3` by `knots = 0:2` in examples